Next set of changes after self testing
Added security configs
Added static .idea configs
Fixed some issues with db integration

+++ spring-server/src/main/java/com/kmoiseev/demo/springserver/view/input/EmployeeInputView.java
@@ -8,5 +8,5 @@ import lombok.Getter;
 @Getter
 public class EmployeeInputView {
     private final String name;
-    private final Integer salary;
+    private final Long salary;
 }
